Split unit and functional tests
